requirements.txt
scrapy
mariadb
python-dotenv